Corrección de una mejora en el código de los ejercicios 4 y 5

+++ 02_Variables/Ejercicios/T02Ejercicio5.java
@@ -7,7 +7,7 @@
 public class T02Ejercicio5 {
   public static void main(String[] args) {
     
-    double peseta = 245;
+    int peseta = 245;
     double euro = peseta * 0.01;
     
     System.out.println("Si tengo " + peseta +" pesetas, a su vez tengo "+ euro+" €");
